% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_drinking_water_source_cat.R
\name{add_drinking_water_source_cat}
\alias{add_drinking_water_source_cat}
\title{Add Drinking Water Source Categories to a Dataset}
\usage{
add_drinking_water_source_cat(
  .dataset,
  drinking_water_source_col = "wash_water_source",
  drinking_water_source_cat_improved_val = c("piped_dwelling", "piped_compound",
    "piped_neighbour", "tap", "borehole", "protected_well", "protected_spring",
    "rainwater_collection", "tank_truck", "cart_tank", "kiosk", "bottled_water",
    "sachet_water"),
  drinking_water_source_cat_unimproved_val = c("unprotected_well", "unprotected_spring"),
  drinking_water_source_cat_surface_water_val = "surface_water",
  drinking_water_source_cat_undefined_val = c("dnk", "pnta", "other")
)
}
\arguments{
\item{.dataset}{A data frame or tibble containing the input data.}

\item{drinking_water_source_col}{The column name for the drinking water source variable
in the dataset. Defaults to `"wash_water_source"`.}

\item{drinking_water_source_cat_improved_val}{A character vector of values classified as
"Improved" drinking water sources. Defaults to:
\describe{
  \item{Includes:}{
    "piped_dwelling", "piped_compound", "piped_neighbour",
    "tap", "borehole", "protected_well", "protected_spring",
    "rainwater_collection", "tank_truck", "cart_tank",
    "kiosk", "bottled_water", "sachet_water"
  }
}}

\item{drinking_water_source_cat_unimproved_val}{A character vector of values classified as
"Unimproved" drinking water sources. Defaults to:
\describe{
  \item{Includes:}{
    "unprotected_well", "unprotected_spring"
  }
}}

\item{drinking_water_source_cat_surface_water_val}{A character value classified as
"Surface Water". Defaults to `"surface_water"`.}

\item{drinking_water_source_cat_undefined_val}{A character vector of values classified as
"Undefined". Defaults to:
\describe{
  \item{Includes:}{
    "dnk" (do not know), "pnta" (prefer not to answer), "other"
  }
}}
}
\value{
A data frame or tibble with an additional column `wash_drinking_water_source_cat`.
}
\description{
This function categorizes drinking water sources into the following categories:
"Improved", "Unimproved", "Surface Water", and "Undefined". It recodes the
drinking water source variable based on provided classifications and adds a new
column `wash_drinking_water_source_cat` to the dataset.
}
\details{
The function performs the following steps:
- Validates that the `.dataset` is a valid data frame or tibble and is not empty.
- Verifies that the `drinking_water_source_col` column exists in the dataset and contains valid responses.
- Categorizes the values in the `drinking_water_source_col` column based on the provided classifications.
- Adds a new column `wash_drinking_water_source_cat` to the dataset with the recoded categories:
  \describe{
    \item{Improved}{Corresponding to entries from `drinking_water_source_cat_improved_val`.}
    \item{Unimproved}{Corresponding to entries from `drinking_water_source_cat_unimproved_val`.}
    \item{Surface Water}{Matching `drinking_water_source_cat_surface_water_val`.}
    \item{Undefined}{Matching `drinking_water_source_cat_undefined_val`.}
  }

If the output column `wash_drinking_water_source_cat` already exists, a warning is issued, and
the column is overwritten.
}
\examples{
# Example dataset
example_data <- data.frame(
  wash_drinking_water_source = c("piped_dwelling", "unprotected_well", "surface_water", "dnk")
)

# Add drinking water source categories
result <- add_drinking_water_source_cat(
  .dataset = example_data,
  drinking_water_source_col = "wash_drinking_water_source"
)

}
