% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_health_barriers.R
\name{add_health_barriers}
\alias{add_health_barriers}
\title{Add Health Barrier Indicators}
\usage{
add_health_barriers(
  .dataset,
  health_barriers_col,
  physical_access_barriers_val,
  financial_access_barriers_val,
  safety_access_barriers_val,
  quality_barriers_val,
  healthcare_seeking_barriers_val,
  availability_barriers_val,
  other_barriers_val,
  no_barriers_val,
  did_not_need_val = NULL
)
}
\arguments{
\item{.dataset}{A data frame or tibble containing the relevant column.}

\item{health_barriers_col}{A character string indicating the column name containing health barriers.}

\item{physical_access_barriers_val}{A character vector of values indicating "physical access barriers."}

\item{financial_access_barriers_val}{A character vector of values indicating "financial access barriers."}

\item{safety_access_barriers_val}{A character vector of values indicating "safety access barriers."}

\item{quality_barriers_val}{A character vector of values indicating "quality-related barriers."}

\item{healthcare_seeking_barriers_val}{A character vector of values indicating "healthcare-seeking behavior barriers."}

\item{availability_barriers_val}{A character vector of values indicating "availability barriers."}

\item{other_barriers_val}{A character vector of values indicating "other barriers."}

\item{no_barriers_val}{A character vector of values indicating "no barriers."}

\item{did_not_need_val}{A character vector of values indicating "did not need" health services.}
}
\value{
A data frame with the following new binary columns:
* **health_barrier_any.physical**
* **health_barrier_any.financial**
* **health_barrier_any.safety**
* **health_barrier_any.quality**
* **health_barrier_any.healthcare_seeking**
* **health_barrier_any.availability**
* **health_barrier_any.other**
* **health_barrier_any.none**
}
\description{
This function identifies specific categories of health barriers from a dataset's column containing select multiple responses, and it creates binary indicator columns for each barrier type.
}
\details{
The input column (`health_barriers_col`) should contain multiple responses in a single cell, separated by delimiters (e.g., commas, spaces). The function searches for specific values corresponding to various barrier categories using `grepl` and creates the following binary (0/1) columns:
* **health_barrier_any.physical**
* **health_barrier_any.financial**
* **health_barrier_any.safety**
* **health_barrier_any.quality**
* **health_barrier_any.healthcare_seeking**
* **health_barrier_any.availability**
* **health_barrier_any.other**
* **health_barrier_any.none**

These columns are coded as `1` if the corresponding values are found in the specified record or `0` if not.
}
\examples{
df <- data.frame(
  health_barriers = c(
    "long distance,cost",
    "poor quality,unsafe facilities",
    "no healthcare available",
    "no barriers",
    "other"
  )
)

df_result <- add_health_barriers(
  .dataset = df,
  health_barriers_col = "health_barriers",
  physical_access_barriers_val = c("long distance", "transport issues"),
  financial_access_barriers_val = c("cost", "expensive"),
  safety_access_barriers_val = c("unsafe facilities", "insecurity"),
  quality_barriers_val = c("poor quality", "lack of equipment"),
  healthcare_seeking_barriers_val = c("refusal", "cultural reasons"),
  availability_barriers_val = c("no healthcare available", "services unavailable"),
  other_barriers_val = c("other"),
  no_barriers_val = c("no barriers")
)

}
