% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hhs.R
\name{add_hhs}
\alias{add_hhs}
\title{Calculate Household Hunger Scale (HHS)}
\usage{
add_hhs(
  .dataset,
  fsl_hhs_nofoodhh = "fsl_hhs_nofoodhh",
  fsl_hhs_nofoodhh_freq = "fsl_hhs_nofoodhh_freq",
  fsl_hhs_sleephungry = "fsl_hhs_sleephungry",
  fsl_hhs_sleephungry_freq = "fsl_hhs_sleephungry_freq",
  fsl_hhs_alldaynight = "fsl_hhs_alldaynight",
  fsl_hhs_alldaynight_freq = "fsl_hhs_alldaynight_freq",
  yes_answer = "yes",
  no_answer = "no",
  rarely_answer = "rarely",
  sometimes_answer = "sometimes",
  often_answer = "often"
)
}
\arguments{
\item{.dataset}{Data frame containing HHS component variables}

\item{fsl_hhs_nofoodhh}{Character: Column for "no food in household" question}

\item{fsl_hhs_nofoodhh_freq}{Character: Column for frequency of no food}

\item{fsl_hhs_sleephungry}{Character: Column for "went to sleep hungry" question}

\item{fsl_hhs_sleephungry_freq}{Character: Column for frequency of sleep hungry}

\item{fsl_hhs_alldaynight}{Character: Column for "whole day without eating" question}

\item{fsl_hhs_alldaynight_freq}{Character: Column for frequency of all day/night}

\item{yes_answer}{Character: Value representing "yes" response}

\item{no_answer}{Character: Value representing "no" response}

\item{rarely_answer}{Character: Value representing "rarely" frequency}

\item{sometimes_answer}{Character: Value representing "sometimes" frequency}

\item{often_answer}{Character: Value representing "often" frequency}
}
\value{
Data frame with added columns: fsl_hhs_score, fsl_hhs_cat, fsl_hhs_cat_ipc
}
\description{
Calculates HHS score and categories based on three core hunger questions
and their frequency responses.
}
