% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_iycf_bof.R
\name{add_iycf_bof}
\alias{add_iycf_bof}
\title{Calculate IYCF Indicator 16: Bottle Feeding (BoF)}
\usage{
add_iycf_bof(
  .dataset,
  iycf_5 = "iycf_5",
  age_months = "age_months",
  yes_val = "yes",
  no_val = "no",
  dnk_val = "dnk"
)
}
\arguments{
\item{.dataset}{Data frame containing IYCF component variables}

\item{iycf_5}{Character: Column name for bottle feeding question (1 = yes)}

\item{age_months}{Character: Column name for child's age in months}

\item{yes_val}{Character: Value representing "yes" in categorical columns}

\item{no_val}{Character: Value representing "no" in categorical columns}

\item{dnk_val}{Character: Value representing "don't know" in categorical columns (recoded to NA)}
}
\value{
Data frame with added column: iycf_bof (1 = bottle fed, 0 = not, NA = ineligible)
}
\description{
Calculates whether a child aged 0-23 months was fed using a bottle with
a nipple in the previous day.
}
