% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_iycf_eibf.R
\name{add_iycf_eibf}
\alias{add_iycf_eibf}
\title{Calculate IYCF Indicator 2: Early Initiation of Breastfeeding (EIBF)}
\usage{
add_iycf_eibf(
  .dataset,
  iycf_2 = "iycf_2",
  age_months = "age_months",
  yes_val = "yes",
  no_val = "no",
  dnk_val = "dnk"
)
}
\arguments{
\item{.dataset}{Data frame containing IYCF component variables}

\item{iycf_2}{Character: Column name for timing of first breastfeed (1 = immediately/within 1 hour, 2 = within 1 hour)}

\item{age_months}{Character: Column name for child's age in months}

\item{yes_val}{Character: Value representing "yes" in categorical columns}

\item{no_val}{Character: Value representing "no" in categorical columns}

\item{dnk_val}{Character: Value representing "don't know" in categorical columns (recoded to NA)}
}
\value{
Data frame with added column: iycf_eibf (1 = early initiation, 0 = not, NA = ineligible)
}
\description{
Calculates whether breastfeeding was initiated within one hour of birth.
Applicable to children 0-23 months.
}
