% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_iycf_evbf.R
\name{add_iycf_evbf}
\alias{add_iycf_evbf}
\title{Calculate IYCF Indicator 1: Ever Breastfed (EvBF)}
\usage{
add_iycf_evbf(
  .dataset,
  iycf_1 = "iycf_1",
  age_months = "age_months",
  yes_val = "yes",
  no_val = "no",
  dnk_val = "dnk"
)
}
\arguments{
\item{.dataset}{Data frame containing IYCF component variables}

\item{iycf_1}{Character: Column name for ever breastfed question (1 = yes)}

\item{age_months}{Character: Column name for child's age in months}

\item{yes_val}{Character: Value representing "yes" in categorical columns}

\item{no_val}{Character: Value representing "no" in categorical columns}

\item{dnk_val}{Character: Value representing "don't know" in categorical columns (recoded to NA)}
}
\value{
Data frame with added column: iycf_evbf (1 = ever breastfed, 0 = not, NA = ineligible)
}
\description{
Calculates whether a child was ever breastfed. Applicable to children 0-23 months.
}
