% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_iycf_mdd.R
\name{add_iycf_mdd}
\alias{add_iycf_mdd}
\title{Calculate IYCF Indicator 8: Minimum Dietary Diversity (MDD)}
\usage{
add_iycf_mdd(
  .dataset,
  age_months = "age_months",
  iycf_4 = "iycf_4",
  iycf_6b = "iycf_6b",
  iycf_6c = "iycf_6c",
  iycf_6d = "iycf_6d",
  iycf_7a = "iycf_7a",
  iycf_7b = "iycf_7b",
  iycf_7c = "iycf_7c",
  iycf_7d = "iycf_7d",
  iycf_7e = "iycf_7e",
  iycf_7f = "iycf_7f",
  iycf_7g = "iycf_7g",
  iycf_7h = "iycf_7h",
  iycf_7i = "iycf_7i",
  iycf_7j = "iycf_7j",
  iycf_7k = "iycf_7k",
  iycf_7l = "iycf_7l",
  iycf_7m = "iycf_7m",
  iycf_7n = "iycf_7n",
  iycf_7o = "iycf_7o",
  yes_val = "yes",
  no_val = "no",
  dnk_val = "dnk"
)
}
\arguments{
\item{.dataset}{Data frame containing IYCF component variables}

\item{age_months}{Character: Column name for child's age in months}

\item{iycf_4}{Character: Column name for currently breastfeeding (1 = yes)}

\item{iycf_6b}{Character: Column name for infant formula}

\item{iycf_6c}{Character: Column name for milk from animals}

\item{iycf_6d}{Character: Column name for yogurt drink}

\item{iycf_7a}{Character: Column name for yogurt food}

\item{iycf_7b}{Character: Column name for grains/roots/tubers}

\item{iycf_7c}{Character: Column name for vitamin A rich vegetables}

\item{iycf_7d}{Character: Column name for white potatoes/tubers}

\item{iycf_7e}{Character: Column name for dark green leafy vegetables}

\item{iycf_7f}{Character: Column name for other vegetables}

\item{iycf_7g}{Character: Column name for vitamin A rich fruits}

\item{iycf_7h}{Character: Column name for other fruits}

\item{iycf_7i}{Character: Column name for organ meat}

\item{iycf_7j}{Character: Column name for flesh meat}

\item{iycf_7k}{Character: Column name for fish/seafood}

\item{iycf_7l}{Character: Column name for eggs}

\item{iycf_7m}{Character: Column name for insects/grubs}

\item{iycf_7n}{Character: Column name for legumes/nuts/seeds}

\item{iycf_7o}{Character: Column name for cheese/other dairy}

\item{yes_val}{Character: Value representing "yes" in categorical columns}

\item{no_val}{Character: Value representing "no" in categorical columns}

\item{dnk_val}{Character: Value representing "don't know" in categorical columns (recoded to NA)}
}
\value{
Data frame with added columns: iycf_mdd_score, iycf_mdd_cat
}
\description{
Calculates the Minimum Dietary Diversity indicator for children 6-23 months.
MDD is met when a child consumes foods from 5 or more of 8 defined food groups.
}
