% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_iycf_mixmf.R
\name{add_iycf_mixmf}
\alias{add_iycf_mixmf}
\title{Calculate IYCF Indicator 5: Mixed Milk Feeding (MixMF)}
\usage{
add_iycf_mixmf(
  .dataset,
  iycf_4 = "iycf_4",
  iycf_6b = "iycf_6b",
  iycf_6c = "iycf_6c",
  age_months = "age_months",
  yes_val = "yes",
  no_val = "no",
  dnk_val = "dnk"
)
}
\arguments{
\item{.dataset}{Data frame containing IYCF component variables}

\item{iycf_4}{Character: Column name for currently breastfeeding (1 = yes)}

\item{iycf_6b}{Character: Column name for infant formula (1 = yes)}

\item{iycf_6c}{Character: Column name for milk from animals (1 = yes)}

\item{age_months}{Character: Column name for child's age in months}

\item{yes_val}{Character: Value representing "yes" in categorical columns}

\item{no_val}{Character: Value representing "no" in categorical columns}

\item{dnk_val}{Character: Value representing "don't know" in categorical columns (recoded to NA)}
}
\value{
Data frame with added column: iycf_mixmf (1 = mixed milk feeding, 0 = not, NA = ineligible)
}
\description{
Calculates whether a breastfed child aged 0-5 months is also receiving
other milk (formula or animal milk). Applicable to children 0-5 months.
}
