% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_iycf_mmf.R
\name{add_iycf_mmf}
\alias{add_iycf_mmf}
\title{Calculate IYCF Indicator 9: Minimum Meal Frequency (MMF)}
\usage{
add_iycf_mmf(
  .dataset,
  age_months = "age_months",
  iycf_4 = "iycf_4",
  iycf_6b_num = "iycf_6b_num",
  iycf_6c_num = "iycf_6c_num",
  iycf_6d_num = "iycf_6d_num",
  iycf_8 = "iycf_8",
  yes_val = "yes",
  no_val = "no",
  dnk_val = "dnk"
)
}
\arguments{
\item{.dataset}{Data frame containing IYCF component variables}

\item{age_months}{Character: Column name for child's age in months}

\item{iycf_4}{Character: Column name for currently breastfeeding (1 = yes)}

\item{iycf_6b_num}{Character: Column name for number of times infant formula was given}

\item{iycf_6c_num}{Character: Column name for number of times milk was given}

\item{iycf_6d_num}{Character: Column name for number of times yogurt was given}

\item{iycf_8}{Character: Column name for number of times solid/semi-solid foods were given}

\item{yes_val}{Character: Value representing "yes" in categorical columns}

\item{no_val}{Character: Value representing "no" in categorical columns}

\item{dnk_val}{Character: Value representing "don't know" in categorical columns (recoded to NA)}
}
\value{
Data frame with added column: iycf_mmf (1 = meets MMF, 0 = not, NA = ineligible)
}
\description{
Calculates the Minimum Meal Frequency indicator for children 6-23 months.
Criteria differ for breastfed vs non-breastfed children and by age group.
}
