% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_iycf_mmff.R
\name{add_iycf_mmff}
\alias{add_iycf_mmff}
\title{Calculate IYCF Indicator 10: Minimum Milk Feeding Frequency for Non-Breastfed Children (MMFF)}
\usage{
add_iycf_mmff(
  .dataset,
  age_months = "age_months",
  iycf_4 = "iycf_4",
  iycf_6b_num = "iycf_6b_num",
  iycf_6c_num = "iycf_6c_num",
  iycf_6d_num = "iycf_6d_num",
  iycf_7a_num = "iycf_7a_num",
  yes_val = "yes",
  no_val = "no",
  dnk_val = "dnk"
)
}
\arguments{
\item{.dataset}{Data frame containing IYCF component variables}

\item{age_months}{Character: Column name for child's age in months}

\item{iycf_4}{Character: Column name for currently breastfeeding (1 = yes)}

\item{iycf_6b_num}{Character: Column name for number of times infant formula was given}

\item{iycf_6c_num}{Character: Column name for number of times milk was given}

\item{iycf_6d_num}{Character: Column name for number of times yogurt drink was given}

\item{iycf_7a_num}{Character: Column name for number of times yogurt food was given}

\item{yes_val}{Character: Value representing "yes" in categorical columns}

\item{no_val}{Character: Value representing "no" in categorical columns}

\item{dnk_val}{Character: Value representing "don't know" in categorical columns (recoded to NA)}
}
\value{
Data frame with added column: iycf_mmff (1 = meets MMFF, 0 = not, NA = ineligible)
}
\description{
Calculates whether a non-breastfed child aged 6-23 months receives at least
2 milk feeds per day (formula, animal milk, yogurt, or dairy products).
}
