% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_iycf_swb.R
\name{add_iycf_swb}
\alias{add_iycf_swb}
\title{Calculate IYCF Indicator 13: Sweet Beverage Consumption (SWB)}
\usage{
add_iycf_swb(
  .dataset,
  age_months = "age_months",
  iycf_6c_swt = "iycf_6c_swt",
  iycf_6d_swt = "iycf_6d_swt",
  iycf_6e = "iycf_6e",
  iycf_6f = "iycf_6f",
  iycf_6g = "iycf_6g",
  iycf_6h_swt = "iycf_6h_swt",
  iycf_6j_swt = "iycf_6j_swt",
  yes_val = "yes",
  no_val = "no",
  dnk_val = "dnk"
)
}
\arguments{
\item{.dataset}{Data frame containing IYCF component variables}

\item{age_months}{Character: Column name for child's age in months}

\item{iycf_6c_swt}{Character: Column name for sweetened milk (1 = yes)}

\item{iycf_6d_swt}{Character: Column name for sweetened yogurt drink (1 = yes)}

\item{iycf_6e}{Character: Column name for fruit juice (1 = yes)}

\item{iycf_6f}{Character: Column name for soda/sugary drinks (1 = yes)}

\item{iycf_6g}{Character: Column name for tea/coffee with sugar (1 = yes)}

\item{iycf_6h_swt}{Character: Column name for other sweetened liquids (1 = yes)}

\item{iycf_6j_swt}{Character: Column name for other sweetened drinks (1 = yes)}

\item{yes_val}{Character: Value representing "yes" in categorical columns}

\item{no_val}{Character: Value representing "no" in categorical columns}

\item{dnk_val}{Character: Value representing "don't know" in categorical columns (recoded to NA)}
}
\value{
Data frame with added column: iycf_swb (1 = consumed sweet beverages, 0 = not, NA = ineligible)
}
\description{
Calculates whether a child aged 6-23 months consumed any sweet beverages
in the previous day.
}
