% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hh_check_roster_1_to_many.R
\name{hh_check_roster_1_to_many}
\alias{hh_check_roster_1_to_many}
\title{Check 1:many linkage integrity}
\usage{
hh_check_roster_1_to_many(hh_df, roster_df, uuid_hh, uuid_roster)
}
\arguments{
\item{hh_df}{A data frame containing household-level data.}

\item{roster_df}{A data frame containing roster (individual-level) data.}

\item{uuid_hh}{Character string specifying the UUID column name in the household data frame.}

\item{uuid_roster}{Character string specifying the UUID column name in the roster data frame that links to households.}
}
\description{
Check 1:many linkage integrity
}
