% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hh_check_roster_relationships.R
\name{hh_check_roster_relationships}
\alias{hh_check_roster_relationships}
\title{Check relationship-to-head plausibility}
\usage{
hh_check_roster_relationships(rel_vec)
}
\arguments{
\item{rel_vec}{Character vector of relationship-to-head codes to validate.}
}
\description{
Check relationship-to-head plausibility
}
