% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hh_normalize_adm.R
\name{hh_normalize_adm}
\alias{hh_normalize_adm}
\title{Normalize administrative names}
\usage{
hh_normalize_adm(x)
}
\arguments{
\item{x}{Character vector of administrative unit names to normalize.}
}
\description{
Normalize administrative names
}
