% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hh_snap_coord.R
\name{hh_snap_coord}
\alias{hh_snap_coord}
\title{Snap GPS coordinates to fixed decimal precision}
\usage{
hh_snap_coord(x, digits = 6)
}
\arguments{
\item{x}{Numeric vector of GPS coordinate values to round.}

\item{digits}{Integer specifying the number of decimal places to round to. Default: 6.}
}
\description{
Snap GPS coordinates to fixed decimal precision
}
