% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iycf_recode_yesno.R
\name{iycf_recode_yesno}
\alias{iycf_recode_yesno}
\title{Recode IYCF yes/no/dnk values to numeric}
\usage{
iycf_recode_yesno(x, yes_val = "yes", no_val = "no", dnk_val = "dnk")
}
\arguments{
\item{x}{Vector to recode}

\item{yes_val}{Character: Value representing "yes"}

\item{no_val}{Character: Value representing "no"}

\item{dnk_val}{Character: Value representing "don't know" (recoded to NA)}
}
\value{
Numeric vector with yes=1, no=2, dnk/other=NA
}
\description{
Internal helper that recodes categorical IYCF responses to numeric values.
If the input is already numeric, it is returned as-is. Otherwise, values
matching yes_val are recoded to 1, no_val to 2, dnk_val to NA, and
unrecognized values to NA.
}
\keyword{internal}
